requirements.txt
torch
transformers
Pillow
torchvision
opencv-python
tqdm
timm
numpy
gradio
openai-whisper
imageio-ffmpeg
